tiltTransitionRisk options

These options are meant to be used mainly by developers or analysts while testing the code or creating data:

tiltIndicatorAfter.output_co2_footprint:

At product level it outputs licensed column co2_footprint.

At company level it outputs the column co2_avg (average co2_footprint by companies_id).

tiltIndicatorAfter.verbose: Controls verbosity.

Examples

library(readr, warn.conflicts = FALSE)
library(dplyr, warn.conflicts = FALSE)
library(withr)
library(tiltToyData)
library(tiltIndicator)
library(tiltIndicatorAfter)

set.seed(1)

restore <- options(list(
  readr.show_col_types = FALSE,
  tiltIndicatorAfter.verbose = TRUE,
  tiltIndicatorAfter.output_co2_footprint = TRUE
))

companies <- read_csv(toy_emissions_profile_any_companies())
products <- read_csv(toy_emissions_profile_products_ecoinvent())
europages_companies <- read_csv(toy_europages_companies())
ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
isic_name <- read_csv(toy_isic_name())

result <- profile_emissions(
  companies,
  products,
  europages_companies = europages_companies,
  ecoinvent_activities = ecoinvent_activities,
  ecoinvent_europages = ecoinvent_europages,
  isic = isic_name
)

result |>
  unnest_product() |>
  select(matches(c("co2")))
#> # A tibble: 532 × 3
#>    co2_footprint co2e_lower co2e_upper
#>            <dbl>      <dbl>      <dbl>
#>  1          303.       18.8       797.
#>  2          303.     -236.        834.
#>  3          303.       76.3       306.
#>  4          303.      157.        651.
#>  5          303.       20.9       445.
#>  6          303.      281.        683.
#>  7           NA        NA          NA 
#>  8          303.       18.8       797.
#>  9          303.     -236.        834.
#> 10          303.       76.3       306.
#> # ℹ 522 more rows

result |>
  unnest_company() |>
  select(matches(c("co2")))
#> # A tibble: 1,728 × 1
#>    co2_avg
#>      <dbl>
#>  1    303.
#>  2    303.
#>  3    303.
#>  4    303.
#>  5    303.
#>  6    303.
#>  7    303.
#>  8    303.
#>  9    303.
#> 10    303.
#> # ℹ 1,718 more rows